[#11680] refactor(idp-basic): Move config validation into IdpRESTFeature
Validate only when the idp-basic extension is enabled: require basic in
gravitino.authenticators and reject basic with simple. Remove
IdpBasicAuthConfigValidator and GravitinoServer startup check.
